refactor(doctor): extract diagnostic domains
